Refactoring code based on comments.
